pyproject.toml
[tool.poetry]
name = "innotter-micro-service"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.7"
djangorestframework = "^3.14.0"
boto3 = "^1.28.82"
pyjwt = "^2.8.0"
black = "^23.11.0"
isort = "^5.12.0"
pytest = "^7.4.3"
celery = "^5.3.4"
psycopg2 = "^2.9.9"
redis = "^5.0.1"
django-celery-beat = "^2.5.0"
rest-framework-simplejwt = "^0.0.2"
requests = "^2.31.0"
pytest-django = "^4.7.0"
pytest-mock = "^3.12.0"
faker = "^20.1.0"

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
include_trailing_comma = true
extend_skip = [".md", ".json"]
src_paths = ["./app/"]

[tool.black]
line_length = 88
src_paths = ["./app/"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
